requirements.txt
Flask 
Flask-Login 
Flask-SQLAlchemy 
pymysql
python-dotenv